Implementação de Dtos, AutoMapper e algumas Data Annotations

+++ BACK/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
+using ProEventos.Application.Dtos;
 using ProEventos.Domain;
-using SQLitePCL;
 
 namespace ProEventos.API.Controllers;
 
@@ -22,9 +22,25 @@ public class EventosController : ControllerBase
         try
         {
             var eventos = await _eventoService.GetAllEventosAsync(true);
-            if(eventos == null) return NotFound("Nenhum evento encontrado");
+            if(eventos == null) return NoContent();
 
-            return Ok(eventos);
+            var eventosRetorno = new List<EventoDto>();
+
+            foreach (var evento in eventos)
+            {
+                eventosRetorno.Add(new EventoDto(){
+                    Id = evento.Id,
+                    Local = evento.Local,
+                    DataEvento = evento.DataEvento,
+                    Tema = evento.Tema,
+                    QtdPessoas = evento.QtdPessoas,
+                    ImagemUrl = evento.ImagemUrl,
+                    Telefone = evento.Telefone,
+                    Email = evento.Email
+                });
+            }
+
+            return Ok(eventosRetorno);
         }
         catch (Exception ex)
         {
@@ -40,7 +56,7 @@ public class EventosController : ControllerBase
        try
         {
             var evento = await _eventoService.GetEventosByIdAsync(id, true);
-            if(evento == null) return NotFound("Evento por Id não encontrado.");
+            if(evento == null) return NoContent();
 
             return Ok(evento);
         }
@@ -57,7 +73,7 @@ public class EventosController : ControllerBase
        try
         {
             var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-            if(evento == null) return NotFound("Eventos por tema não encontrados.");
+            if(evento == null) return NoContent();
 
             return Ok(evento);
         }
@@ -69,11 +85,11 @@ public class EventosController : ControllerBase
     }
 
     [HttpPost]
-    public async Task<IActionResult> Post(Evento model){
+    public async Task<IActionResult> Post(EventoDto model){
         try
         {
             var evento = await _eventoService.AddEvento(model);
-            if(evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+            if(evento == null) return NoContent();
 
             return Ok(evento);
         }
@@ -85,11 +101,11 @@ public class EventosController : ControllerBase
     }
 
     [HttpPut("{id}")]
-    public async Task<IActionResult> Put(int id, Evento model){
+    public async Task<IActionResult> Put(int id, EventoDto model){
         try
         {
             var evento = await _eventoService.UpdateEvento(id, model);
-            if(evento == null) return BadRequest("Erro ao tentar atualizar o evento.");
+            if(evento == null) return NoContent();
 
             return Ok(evento);
         }
@@ -104,10 +120,13 @@ public class EventosController : ControllerBase
     public async Task<IActionResult> Delete(int id){
         try
         {
+            var evento = await _eventoService.GetEventosByIdAsync(id, true);
+            if (evento == null) return NoContent();
+
             if (await _eventoService.DeleteEvento(id))
                 return Ok("Evento deletado");
             else
-                return BadRequest("Não foi possivel deletar o evento");
+                throw new Exception("Ocorreu um problema não especificado ao tentar deletar o Evento: #" + id);
         }
         catch (Exception ex)
         {
